Moved the control grid hiding logic
* Moved the control grid hiding logic from the videoplayerpageviewmodel to the code behind because its only doing something with th view.

* Removed unused usings and sorted usings in the videoplayerpageviemodel.
