+++ OneDrive-Cloud-Player/Views/VideoPlayerPage.xaml
@@ -12,12 +12,8 @@
     mc:Ignorable="d"
     DataContext="{Binding VideoPlayerPageInstance, Source={StaticResource Locator}}"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-    <Grid>
-        <Interactivity:Interaction.Behaviors>
-            <Core:EventTriggerBehavior EventName="PointerMoved">
-                <Core:InvokeCommandAction Command="{Binding PointerMovedMediaPlayerCommand}"/>
-            </Core:EventTriggerBehavior>
-        </Interactivity:Interaction.Behaviors>
+    
+    <Grid PointerMoved="Grid_PointerMoved">
         <lvs:VideoView MediaPlayer="{Binding MediaPlayer}">
             <Interactivity:Interaction.Behaviors>
                 <Core:EventTriggerBehavior EventName="Initialized">
@@ -27,16 +23,9 @@
         </lvs:VideoView>
 
         <Grid Background="#1E2931" x:Name="MediaControlGrid" MaxHeight="200"
-              VerticalAlignment="Bottom" Visibility="{Binding MediaControlGridVisibility}">
-
-            <Interactivity:Interaction.Behaviors>
-                <Core:EventTriggerBehavior EventName="PointerEntered">
-                    <Core:InvokeCommandAction Command="{Binding PointerEnteredMediaControlGridCommand}"/>
-                </Core:EventTriggerBehavior>
-                <Core:EventTriggerBehavior EventName="PointerExited">
-                    <Core:InvokeCommandAction Command="{Binding PointerExitedMediaControlGridCommand}"/>
-                </Core:EventTriggerBehavior>
-            </Interactivity:Interaction.Behaviors>
+              VerticalAlignment="Bottom" Visibility="{Binding MediaControlGridVisibility}"
+              PointerEntered="MediaControlGrid_PointerEntered"
+              PointerExited="MediaControlGrid_PointerExited">
 
             <Grid  x:Name="LowerHeightControls">
                 <Grid.ColumnDefinitions>
