+++ OneDrive-Cloud-Player/Views/VideoPlayerPage.xaml.cs
@@ -7,6 +7,9 @@ using System.Linq;
 using System.Runtime.InteropServices.WindowsRuntime;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
+using Windows.UI.Core.Preview;
+using Windows.UI.Popups;
+using Windows.UI.ViewManagement;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Controls.Primitives;
@@ -24,10 +27,22 @@ namespace OneDrive_Cloud_Player.Views
     /// </summary>
     public sealed partial class VideoPlayerPage : Page
     {
+        private readonly DispatcherTimer pointerMovementDispatcherTimer;
+        private bool isPointerOverMediaControlGrid;
+
         public VideoPlayerPage()
         {
             this.InitializeComponent();
+
             SeekBar.AddHandler(PointerPressedEvent, new PointerEventHandler(SeekBar_PointerPressed), true);
+
+            //Create a timer with interval.
+            pointerMovementDispatcherTimer = new DispatcherTimer();
+            pointerMovementDispatcherTimer.Tick += PointerMovementDispatcherTimer_Tick;
+
+            pointerMovementDispatcherTimer.Interval = new TimeSpan(0, 0, 2);
+
+            pointerMovementDispatcherTimer.Start();
         }
 
         /// <summary>
@@ -60,7 +75,54 @@ namespace OneDrive_Cloud_Player.Views
             {
                 throw;
             }
-            
+
+        }
+
+        /// <summary>
+        /// Gets called when a pointer enters the media control grid of the mediaplayer.
+        /// </summary>
+        private void MediaControlGrid_PointerEntered(object sender, PointerRoutedEventArgs e)
+        {
+            isPointerOverMediaControlGrid = true;
+            Debug.WriteLine(" + Pointer entered control grid.");
+            //MediaControlGridVisibility = "Collapsed";
+
+        }
+
+        /// <summary>
+        /// Gets called when a pointer exits the media control grid of the mediaplayer.
+        /// </summary>
+        private void MediaControlGrid_PointerExited(object sender, PointerRoutedEventArgs e)
+        {
+            isPointerOverMediaControlGrid = false;
+            Debug.WriteLine(" + Pointer exited control grid.");
+        }
+
+        /// <summary>
+        /// Gets called when a pointer moves across the mediaplayer.
+        /// </summary>
+        private void Grid_PointerMoved(object sender, PointerRoutedEventArgs e)
+        {
+            if (MediaControlGrid.Visibility == Visibility.Collapsed)
+            {
+                MediaControlGrid.Visibility = Visibility.Visible;
+            }
+            pointerMovementDispatcherTimer.Start();
+        }
+
+        /// <summary>
+        /// Gets called when the dispatcher event fires of the pointer movement.
+        /// It collapses the mediaplayer control grid on certain conditions.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void PointerMovementDispatcherTimer_Tick(object sender, object e)
+        {
+            if (!isPointerOverMediaControlGrid)
+            {
+                MediaControlGrid.Visibility = Visibility.Collapsed;
+            }
+            pointerMovementDispatcherTimer.Stop();
         }
     }
 }
